pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[project]
name = "capsulitis"
version = "25.1b0"
authors = [
    { name="Tsz Kiu Pang" },
]
description = "Capsulitis"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "abjad @ git+https://github.com/Abjad/abjad.git@7bca4dc7ab9b55441c3196e6e6d5b9e9184cc36f",
    "abjad-ext-nauert @ git+https://github.com/Abjad/abjad-ext-nauert.git@8dec05b66f537fba9ec153742658c05694c3aaec",
    "pang",
    "numpy >= 2.0"
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]

[project.optional-dependencies]
test = [
    "black",
    "flake8",
    "isort",
    "mypy",
    "pytest",
]

[tool.hatch.build.targets.wheel]
packages = ["capsulitis"]
